Changed old photos migrator
