+++ app/src/main/java/com/fieldbook/tracker/traits/PhotoTraitLayout.kt
@@ -22,9 +22,6 @@ import com.fieldbook.tracker.R
 import com.fieldbook.tracker.activities.CollectActivity
 import com.fieldbook.tracker.activities.ConfigActivity
 import com.fieldbook.tracker.adapters.GalleryImageAdapter
-import com.fieldbook.tracker.database.dao.ObservationDao
-import com.fieldbook.tracker.database.dao.ObservationVariableDao
-import com.fieldbook.tracker.database.models.ObservationModel
 import com.fieldbook.tracker.objects.TraitObject
 import com.fieldbook.tracker.preferences.GeneralKeys
 import com.fieldbook.tracker.utilities.DialogUtils
@@ -79,13 +76,6 @@ class PhotoTraitLayout : BaseTraitLayout {
         capture.setOnClickListener(PhotoTraitOnClickListener())
         photo = findViewById(R.id.photo)
         activity = act
-        scope = CoroutineScope(Dispatchers.IO)
-
-        scope.launch {
-
-            migrateOldPhotosDir()
-
-        }
     }
 
     override fun loadLayout() {
@@ -96,76 +86,6 @@ class PhotoTraitLayout : BaseTraitLayout {
         loadLayoutWork()
 
         super.loadLayout()
-    }
-
-    /**
-     * In v5.3 a directory change happened, where each trait gets it's own media directory for photos.
-     * This function is called to query for existing observations that have a uri in the old photos directory, which used to hold all photos.
-     * This will update the database obs. value to a new uri after copying it to its respective trait folder, and delete the old photo from the photos dir.
-     */
-    private fun migrateOldPhotosDir() {
-
-        try {
-
-            ObservationVariableDao.getAllTraitObjects().filter { it.format == type }.forEach { t ->
-
-                val timeStamp = SimpleDateFormat(
-                    "yyyy-MM-dd-hh-mm-ss", Locale.getDefault()
-                )
-
-                val expId = prefs.getInt(GeneralKeys.SELECTED_FIELD_ID, 0).toString()
-
-                val traitPhotos = ObservationDao.getAll(expId).filter { it.observation_variable_name == t.trait }
-
-                if (t.trait != "photos") { //edge case where trait name is actually photos
-
-                    val photoDir = getFieldMediaDirectory(context, t.trait)
-                    val oldPhotos = getFieldMediaDirectory(context, "photos")
-
-                    traitPhotos.forEach { photo ->
-
-                        //todo update rep
-                        val repeatedValue = ConfigActivity.dt.getRep(photo.observation_unit_id, t.trait)
-                        val generatedName =
-                            photo.observation_unit_id + "_" + t.trait + "_" + repeatedValue + "_" + timeStamp.format(
-                                Calendar.getInstance().time
-                            ) + ".jpg"
-
-                        //load uri and check if its parent is "photos" old photo dir
-                        oldPhotos?.findFile(photo.value ?: "")?.let { photoFile ->
-
-                            photoDir?.createFile("*/jpg", photoFile.name ?: generatedName)?.let { newFile ->
-
-                                context.contentResolver?.openInputStream(photoFile.uri)?.use { input ->
-
-                                    context.contentResolver?.openOutputStream(newFile.uri)?.use { output ->
-
-                                        input.copyTo(output)
-
-                                    }
-                                }
-
-                                ObservationDao.updateObservation(ObservationModel(
-                                    photo.createMap().apply {
-                                        this["value"] = newFile.uri.toString()
-                                    }
-                                ))
-
-                                createThumbnail(photoDir, newFile.uri)
-                            }
-
-                            photoFile.delete()
-
-                        }
-                    }
-                }
-            }
-
-        } catch (e: Exception) {
-
-            Log.e(TAG, "Error during photo migration", e)
-
-        }
     }
 
     private fun loadLayoutWork() {
